# swagger-example.yaml
openapi: 3.0.3
info:
    title: Swagger Petstore - OpenAPI 3.0
    description: |-
        This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
        Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
        You can now help us improve the API whether it's by making changes to the definition itself or to the code.
        That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

        _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

        Some useful links:
        - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
        - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
    termsOfService: http://swagger.io/terms/
    contact:
        email: apiteam@swagger.io
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 1.0.11
...
servers:
    - url: https://petstore3.swagger.io/api/v3
...
paths:
...
    /user:
        post:
            tags:
                - user
            summary: Create user
            description: This can only be done by the logged in user.
            operationId: createUser
            requestBody:
                description: Created user object
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/User"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/User"
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: "#/components/schemas/User"
            responses:
                default:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/User"
...
components:
    schemas:
...
        User:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                    example: 10
                username:
                    type: string
                    example: theUser
                firstName:
                    type: string
                    example: John
                lastName:
                    type: string
                    example: James
                email:
                    type: string
                    example: john@email.com
                password:
                    type: string
                    example: "12345"
                phone:
                    type: string
                    example: "12345"
                userStatus:
                    type: integer
                    description: User Status
                    format: int32
                    example: 1
            xml:
                name: user
...
    securitySchemes:
        petstore_auth:
            type: oauth2
            flows:
                implicit:
                    authorizationUrl: https://petstore3.swagger.io/oauth/authorize
                    scopes:
                        write:pets: modify pets in your account
                        read:pets: read your pets
        api_key:
            type: apiKey
            name: api_key
            in: header